content_main.cc: Initialize `JSPreloadFrameHost` globally
The previous patch adds a new class `base::JSPreloadFrameHost`, we
start using that class in this patch.

Several things that this patch does:

1) Create a global variable `base::JSPreloadFrameHost* g_js_preload`.
   This variable is initialized with a `nullptr` and will be set to a
   new object of that class. The object construction is done in the
   `RunContentProcess()` function.

2) On the object construction, we load a JavaScript file from the
   given file name. The file name is taken from an environment
   variable named `CHROMIUM_JS_PRELOAD_FILE`.

3) Right before the `RunContentProcess()` returns, we `delete` the
   object to prevent memory leak.

Between steps (2) and (3), our internal handler will use the
preloaded JavaScript code that step (2) provides to register a
JavaScript extension (this is done in the next patches).

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>