task_am_001: Call the preloaded JS script in all frames
This will execute the following small JavaScript code in all frames:
```
  if (typeof __preloaded_javascript_func == 'function') {
     __preloaded_javascript_func();
  }
```

The preloaded JavaScript code must live in a function called
`__preloaded_javascript_func()`, this function will be called in all
frames before the page loads. We can't execute the code directly from an
extension registration, so we do it here (in the render_frame_impl.cc).

This means, to set a preloaded JavaScript code, we must write it like
this (this code example will call the alert() in all frames):
```
  function __preloaded_javascript_func() {
    alert("This is an alert from the JavaScript preloaded code");
  }
```

Extra note for modifiying HTML elements. We need to wait until they
are properly loaded. For example, the following code won't work and
results in undefined (even if there is a HTML tag with ID "test"):
```
  function __preloaded_javascript_func() {
    document.getElementById("test").innerHTML = "Hello";
  }
```

This is because the code is executed before even the page is loaded.
As such, we need to register a callback to window onload event like
this:
```
  function __preloaded_javascript_func() {
    window.addEventListener("load", function () {
      document.getElementById("test").innerHTML = "Hello";
    });
  }
```

Extra note again, don't use `window.onload = function () {}`, because
it can only handle a single callback. If the page loads a JavaScript
code that override `window.onload`, our preloaded script won't run.

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>