toolbar/toolbar_view: Hide address bar, buttons, menu, etc.
This patch hides many things on the toolbar. Currently there are 16 items to be
hidden from there:
  01) left_side_panel_button_
  02) back_
  03) forward_
  04) reload_
  05) home_
  06) custom_tab_bar_
  07) location_bar_
  08) extensions_container_
  09) chrome_labs_button_
  10) cast_
  11) side_panel_button_
  12) toolbar_account_icon_container_
  13) avatar_
  14) media_button_
  15) send_tab_to_self_button_
  16) app_menu_button_

@@ Short technical explanation:
To hide these view components, we can call `SetVisible(false)`. For example,
hiding home button can simply be: `home_->SetVisible(false)`.

@@ Note for Developers (including future developers who maintain this patch):
The above item list may change over time. To maintain it, you can take a look at
file: chrome/browser/ui/views/toolbar/toolbar_view.cc

That file is very unlikely to be deleted, but it is very likely to change. So
that file is a stable hint where to maintain this patch functionality.

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>