render_frame_impl.cc: Add native JavaScript array processing
The second task is about creating a native function that can take
arrays from JavaScript in order to process them on the native side.

This is an example of a native function that will be accessible in
JavaScript. The function can take 2 arguments an Array and a Number,
respectively. Only the first argument is required, as the second
argument defaults to the value 1.

This example shows a function that takes any size array of numbers
and multiplies them by the second argument.

1. Compile & Run chromium
2. Open: DevTools > Console
3. Run the following snippet:
```js
let x = [1, 2, 3, 4, 5, 6, 7];
mul(x, 3);
console.log(x); // 3, 6, 9, 12, 15, 18, 21
```

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>