task_am_001: base: Create `JSPreloadFrameHost` class
This is used to preload a JavaScript file at initialization. It reads
JavaScript code from the given filename and saves it on a global
state. The saved JavaScript code will be reused as much as possible.
The only thing that requires us to re-open the file is that when we
have a separate process that doesn't have the `g_js_preload` yet (In
other words, when `g_js_preload` is still nullptr).

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>