task_am_002: Add --hide-all command line argument
This adds --hide-all command line argumnet for task_am_002. The task
is to hide everything like tabs, settings, menu bar, title bar, etc.

In short: Hide everything, except the content frame.

This commit is the preparation for that, it only applies --hide-all
argument to all renderer processes when this argument is provided.
We will do the hide part in the next patches.

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>