task_g_028: Add x-hdy headers (for all other requests, probably)
I don't really know what this part is actually doing, but according
to the patch instruction I received, this is to add x-hdy headers for
all requests (probably).

The instruction I received is also not sure, they put "probably" word.

> All requests (probably)

Let's see if I can manage to get it working...

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>