task_g_028: Fix missing x-hdy header for websocket connection from an extension
Regarding the missing header issue, after several hours playing with
it, I figured out that for some particular cases (e.g. websocket from
an extension). The chromium doesn't call
`WebSocketConnectorImpl::Connect()`, instead it calls
`WebSocketConnectorImpl::ConnectCalledByContentBrowserClient()`.

The latter is a static function, so we can't just simply take the
`std::vector` of X-Hdy headers that lives in the
`WebSocketConnectorImpl` class.

Fix this by appending the X-Hdy-Browser-Id header directly from the
`WebSocketConnectorImpl::ConnectCalledByContentBrowserClient()`
function. At that point, we may not have frame ID, host, etc. As
the request may not be executed from a general user context, like
for example, when accessing a general URL directly.

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>