app/chrome_main: Create non static version of xhdy helpers
If we can use the non-static function, use the non-static function.

These functions are "static inline":
  - __is_hdy_headers_on()
  - __get_browser_id()

These function are global and should be preferred to use:
  - is_hdy_headers_on()
  - get_browser_id()

The reason of why we have the static and non-static function is that,
not all area of the chromium can see those symbols, in case it can't,
use the static inline, otherwise, use the non static version.

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>