url_request_http_job: Provide non-static xhdy helpers for net
Always try to use the non-static function first, if we fail during the
linking time, that means the symbols are not visible from the compiled
object. In that case, provide the non-static definition and tail-call
them to the static inline definition.

Do the same thing on the net process now.

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>