task_g_028: Fix missing x-hdy headers on websocket when using extension
When using an extension, the websocket goes in different path because
at that point the WillInterceptWebSocket() method in the file
websocket_connector_impl.cc returns true, so it's somewhat proxied
to a different function.

I don't fully understand this part yet, but killing that path fixes
the problem. We can revert this change if it's problematic, but so
far I don't see any problem.

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>